turn CSVUtility into a full static utility-class using @UtilityClass annotation fix toCsvFile method header issue
